完成数据分析

+++ server/model/request/stock_detail.go
@@ -1,8 +1,8 @@
 package request
 
 type StockTopListReq struct {
-	MarketCapitalMax int64   `json:"marketCapitalMax" desc:"市值范围最大" `
-	MarketCapitalMin int64   `json:"marketCapitalMin" desc:"市值范围最小" `
+	MarketCapitalMax float64 `json:"marketCapitalMax" desc:"市值范围最大" `
+	MarketCapitalMin float64 `json:"marketCapitalMin" desc:"市值范围最小" `
 	PercentMax       float64 `json:"percentMax" desc:"涨幅范围最大" `
 	PercentMin       float64 `json:"percentMin" desc:"涨幅范围最小" `
 	VolumeRatioMax   float64 `json:"volume_ratio_max" desc:"量比范围最大" `
@@ -15,8 +15,8 @@ type AllListReq struct {
 	PageInfo
 	Name             string  `json:"name" desc:"名称" `
 	Code             string  `json:"code" desc:"编码" `
-	MarketCapitalMax int64   `json:"marketCapitalMax" desc:"市值范围最大" `
-	MarketCapitalMin int64   `json:"marketCapitalMin" desc:"市值范围最小" `
+	MarketCapitalMax float64 `json:"marketCapitalMax" desc:"市值范围最大" `
+	MarketCapitalMin float64 `json:"marketCapitalMin" desc:"市值范围最小" `
 	PercentMax       float64 `json:"percentMax" desc:"涨幅范围最大" `
 	PercentMin       float64 `json:"percentMin" desc:"涨幅范围最小" `
 	VolumeRatioMax   float64 `json:"volume_ratio_max" desc:"量比范围最大" `
@@ -33,8 +33,8 @@ type MonitorListReq struct {
 	IsDay            bool    `json:"isDay" desc:"是否日监控" `
 	Name             string  `json:"name" desc:"名称" `
 	Code             string  `json:"code" desc:"编码" `
-	MarketCapitalMax int64   `json:"marketCapitalMax" desc:"市值范围最大" `
-	MarketCapitalMin int64   `json:"marketCapitalMin" desc:"市值范围最小" `
+	MarketCapitalMax float64 `json:"marketCapitalMax" desc:"市值范围最大" `
+	MarketCapitalMin float64 `json:"marketCapitalMin" desc:"市值范围最小" `
 	PercentMax       float64 `json:"percentMax" desc:"涨幅范围最大" `
 	PercentMin       float64 `json:"percentMin" desc:"涨幅范围最小" `
 	VolumeRatioMax   float64 `json:"volume_ratio_max" desc:"量比范围最大" `
@@ -64,8 +64,8 @@ type AnalysisListReq struct {
 	PageInfo
 	Name             string  `json:"name" desc:"名称" `
 	Code             string  `json:"code" desc:"编码" `
-	MarketCapitalMax int64   `json:"marketCapitalMax" desc:"市值范围最大" `
-	MarketCapitalMin int64   `json:"marketCapitalMin" desc:"市值范围最小" `
+	MarketCapitalMax float64 `json:"marketCapitalMax" desc:"市值范围最大" `
+	MarketCapitalMin float64 `json:"marketCapitalMin" desc:"市值范围最小" `
 	PercentMax       float64 `json:"percentMax" desc:"涨幅范围最大" `
 	PercentMin       float64 `json:"percentMin" desc:"涨幅范围最小" `
 	VolumeRatioMax   float64 `json:"volume_ratio_max" desc:"量比范围最大" `
