+++ server/service/stock_top.go
@@ -14,12 +14,12 @@ func (*stockTopService) TopList(filter request.StockTopListReq) (stockList []mod
 
 	for _, stock := range stockListCrawler {
 		if filter.MarketCapitalMax > 0 {
-			if stock.MarketCapital > filter.MarketCapitalMax {
+			if stock.MarketCapital > int64(filter.MarketCapitalMax) {
 				continue
 			}
 		}
 		if filter.MarketCapitalMin > 0 {
-			if stock.MarketCapital < filter.MarketCapitalMin {
+			if stock.MarketCapital < int64(filter.MarketCapitalMin) {
 				continue
 			}
 		}
